Create index.js
This file is essentially a "barrel" or "index" file that exports the Block and Blockchain modules defined in the src/blockchain/block.js and src/blockchain/blockchain.js files, respectively. This allows other parts of the application to easily import and use these modules by referencing the src/blockchain directory.