Create main.js
This file imports the Blockchain class from the src/blockchain/blockchain.js file and creates an instance of it. Then it adds new blocks to the chain by calling the addBlock() method, passing it the new block data. Finally, it logs the entire chain to the console and checks if the chain is valid by calling the isChainValid() method. This is a basic example of how the main.js file could be used to demonstrate the functionality of the blockchain implementation. In a real-world application, this file would likely contain additional logic such as starting a local server or connecting to other nodes in a peer-to-peer network.

Also, the actual implementation will vary depending on the specific requirements and design of the project and this is just a simple demonstration on how to use the blockchain class.